Fix non-escaped forward slash podchecker warning
POD parsers can have problems when forward slashes appear in the text part
of a POD link (`L<>`).  This change avoids the issue without changing the
sense of the sentence.  If one uses the advice in `Pod::Checker` for
handline this issue, one would have to use `PSGIE<sol>Plack` which is simply
too hard to read, although avoids the error while preserving the forward
slash.  In this case it seems that a hyphen is sufficient to retain the
meaning while avoiding any potential POD parser problems.