Add Moose::Autobox dependency to dist.ini
This dependency needs to be specified as an author dependency since
`SeeAlsoMason.pm` depends upon it, however for some reason `dzil` doesn't
list it as a requirement as part of `dzil authordeps` even if it is listed
as a normal dependency under the `[Prereqs / RuntimeRequires]` section.
Thus, if this module isn't installed, `dzil -Ilib listdeps` will fail and
won't list *any* of the required module dependencies.  This change works
around the issue, listing the module under the author dependencies such that
`dzil -Ilib authordeps | cpanm --no-skip-satisfied` installs the required
modules for the further functioning of the module when installing from the
sources.